+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 0.0.2
+
+* Update Audienzz native dependencies to latest version
+* Remove deprecated parameters 'appContent', 'keyword' and 'keywords'
+
 ## 0.0.1
 
 * Initial release of the sdk wrapper. Supports Banner, Interstitial and Rewarded ad formats.